+++ src/rs/symbol_decomp_state.rs
@@ -54,19 +54,7 @@ impl<'sect> Iterator for SymbolDecompStateIter<'sect> {
         let sym = &self.section.symbols[self.index];
         self.index += 1;
 
-        // TODO: move `.NON_MATCHING` handling to the parsing itself instead.
-        if let Some(non_matching_sym) = self
-            .section
-            .find_symbol_by_name(&format!("{}.NON_MATCHING", sym.name))
-        {
-            let mut undecomped_sym = sym.clone();
-            if undecomped_sym.size == 0 {
-                undecomped_sym.size = non_matching_sym.size;
-            }
-            return Some(SymbolDecompState::Undecomped(undecomped_sym));
-        }
-
-        if self.whole_file_is_undecomped {
+        if self.whole_file_is_undecomped || sym.nonmatching_sym_exists {
             return Some(SymbolDecompState::Undecomped(sym.clone()));
         } else if let Some(functions_path) = &self.functions_path {
             if functions_path.join(sym.name.clone() + ".s").exists() {